Remove is_cwfile.
This function was used for two purposes: first, to check we don't
fork-bomb, but that would require writing a bogus path directive. Not
very likely. Secondly, it was used as a check before trying to parse a
definition file, but that's silly: a valid definition file is one we
can parse. There's no reason in principle why you shouldn't want
non-executable definition files that don't start with a hash-bang
line.