Remove usepty, --cw-usepty and CW_USEPTY.
Always use a pty if we can when coloring is on.

This was the last "support" letter mentioned in the man page. Rename
"support" to "features" and make the features words rather than single
letters (there are still two, pty and setproctitle, which it could be
useful to ask users to report when debugging; but neither is now
user-controllable).