Replace configuration files with Lua scripts.
A function "match" is called to add match patterns. The command-line
arguments are available in the arg table, both in numbered form, and
with a key for each argument, to make it easy to find out whether a
particular argument such as --help exists in any position.

The OS type can be found via posix.uname, and environment variables
can be set with posix.setenv.