cw.c: simplify set_up_harness and sighandler
Remove the eint flag, which is just pid_c!=0

Remove color stripping of the input, as it isn't needed for color
management, and its useful function (removing the printable characters
in control sequences which might accidentally match color patterns) is
not comprehensive, as it only matches colors, not other control
sequences. We shouldn't have to deal with colored input, anyway: if a
program is coloring its output, we shouldn't be wrapping it!